feat(purchase,inventory): drive qty input type from Item.qty_category, not weight field
Cancels the just-shipped "Actual Weight" field on Goods Receipt (wrong
fix) and replaces it with the real one: qty's integer-vs-decimal input
type depends on how the specific item is measured (zak/pcs = whole
number, bulk cement weighed on a truck scale = 2 decimals), not on a
separate weight fact. Item.qty_category ('unit' | 'weight') decides
this end to end.

- Item.qty_category (new column, defaults 'unit' — zero behavior change
  for existing items) with App\Enums\QtyCategory as the single source of
  truth for decimal places (0 / 2).
- Every qty-bearing column widened back to DECIMAL(18,4): PO/GR/Invoice/
  Return qty, Item.current_stock, StockLedger, plus Stock Adjustment and
  Stock Transfer (not touched by the earlier qty-decimal sprint). A
  qty_category snapshot column rides alongside each line-item table, same
  historical-accuracy rationale as their existing `uom` string snapshot.
- New App\Services\QtyCategoryValidator, shared by GoodsReceiptService,
  PurchaseOrderService, PurchaseReturnService, StockAdjustmentService,
  and StockTransferService: rejects a non-whole qty on a Unit-category
  item, rounds a Weight-category item's qty to 2dp before it's stored —
  business-rule validation in the Service layer, not the FormRequest
  (which never has the Item loaded), same pattern as
  assertWithinOutstanding()/assertSufficientStock().
- Frontend: one shared frontend/src/shared/lib/qty.ts (parse comma-or-dot
  input, format per category, category-aware zod validation) used by
  every line-item form and read-only qty display across these 5 modules.
  Each item Select now snapshots the selected item's qty_category onto
  the row and drives the qty input's step/min accordingly.

+++ backend/laravel-api/app/Http/Requests/StorePurchaseReturnRequest.php
@@ -23,7 +23,9 @@ class StorePurchaseReturnRequest extends FormRequest
             'remarks' => ['nullable', 'string'],
             'items' => ['sometimes', 'array'],
             'items.*.purchase_invoice_item_id' => ['required_with:items', 'uuid', 'exists:purchase_invoice_items,id'],
-            'items.*.qty_returned' => ['sometimes', 'integer', 'min:0'],
+            // Whole-number-vs-decimal enforcement happens in PurchaseReturnService via
+            // QtyCategoryValidator (needs the Item loaded, not available here).
+            'items.*.qty_returned' => ['sometimes', 'numeric', 'min:0'],
             'items.*.amount' => ['required_with:items', 'numeric', 'min:0'],
         ];
     }
